+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
 ## 1.9.0
-* Add `is_privilege_in_role?` method to ActiveRecord storage adapter.
+* Add `is_privilege_via_attribute?` method to ActiveRecord storage adapter.
 * Update `accessible_ancestor_objects`, `accessible_objects`, `scoped_privileges`, and `accessible_operations` methods in the ActiveRecord storage adapter to accept scoping to a user attribute.
 * Update `is_privilege_ignoring_prohibitions?` to accept an optional user attribute scope.